feat: Make a simple YjsSharedDocument that doesn't require the network
